💄 style: Update broswer
